[nvim]: Disable scope highlighting for `indent-blankline`, tweak highlight colour
Generally speaking I've found scope highlighting to reduce performance
when scrolling with the mouse wheel, as well as when scrolling line by
line with either j or k.

I've found the indent highlighting to be more valuable than the scope
highlighting recently, so I'm going to disable it for now just to see if
I notice a difference with it off.